+++ functions/index.js
@@ -1,6 +1,5 @@
-exports.brackets = require('./brackets').brackets;
+exports.cards = require('./cards').cards;
 exports.build = require('./build').build;
-exports.card = require('./card').card;
 exports.faq = require('./faq').faq;
 exports.deck = require('./deck').deck;
 exports.errata = require('./errata').errata;