+++ functions/rule.js
@@ -4,7 +4,8 @@ const rules = require('../card_data/rules');
 const _ = require('lodash');
 
 const rule = (msg, params) => {
-	let i = _.findIndex(Object.keys(rules), term => term === params.join(' ') || term.startsWith(params.join(' ')));
+	let i = _.findIndex(Object.keys(rules), term => term === params.join(' '));
+	if (!i) i = _.findIndex(Object.keys(rules), term => term.startsWith(params.join(' ')));
 	const key = Object.keys(rules)[i];
 	let embed = new Discord.RichEmbed();