+++ handlers.js
@@ -7,50 +7,43 @@ const _ = require('lodash');
 // Called every time a message comes in:
 const onMessage = (msg, client) => {
 	if (msg.author.bot) return; // Ignore messages from the bot
-	let params = msg.content.split(' '),
+	let params = msg.content.toLowerCase().split(' '),
 		commandName = params.map(a => a.startsWith(commandPrefix) && a).filter(Boolean).join().slice(1),
 		flags = params.filter(a => a.startsWith('-')).map(flag => flag.slice(1)),
 		types = {
-			brackets: {
-				regex: /\[(.*?)\]/,
-				params: []
-			},
-			d: {
-				regex: /[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}/,
-				params: []
-			},
-			deck: {
-				regex: /\{(.*?)\}/,
-				params: []
-			},
+			cards: /\[(.*?)\]/,
+			d: /[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}/,
+			deck: /\{(.*?)\}/,
 		};
+
 	if (commandName) {
-		if (msg.content.startsWith(commandPrefix)) params = msg.content.substr(1).split(' ');
-		else params = msg.content.split(commandPrefix)[1].split(' ');
-		params = params.splice(1).filter(a => !a.startsWith('-'));
+		params = _.dropWhile(params, a => !a.includes(commandPrefix))
+			.slice(1)
+			.filter(a => !a.startsWith('-'));
+	} else {
+		Object.keys(types).forEach(a => {
+			let message = msg.content.toLowerCase(), arr = [];
+			do {
+				let param = message.match(types[a]);
+				if (param) {
+					arr.push(_.get(param, '1', param[0]));
+					message = message.replace(_.get(param, '0'), '');
+				}
+			} while (message.match(types[a]));
+			if (arr.length > 0) {
+				commandName = a;
+				if (params < 1) flags = [...flags, 'delete'];
+				params = arr;
+			}
+		});
 	}
 
-	Object.keys(types).forEach(a => {
-		let message = msg.content;
-		do {
-			let param = message.match(types[a].regex);
-			if (param) {
-				types[a].params.push(_.get(param, '1', param[0]));
-				message = message.replace(_.get(param, '0'), '');
-			}
-		} while (message.match(types[a].regex));
-		if (types[a].params.length > 0) {
-			commandName = a;
-			if (params < 1) flags = [...flags, 'delete'];
-			params = types[a].params
-		}
-	});
-
 	if (!commandName) return;
-
+	params = params.filter(Boolean);
 	switch (commandName) {
 		case 'c':
-			commandName = 'card';
+		case 'card':
+			commandName = 'cards';
 			break;
 		case 'd':
 			commandName = 'deck';